[FIX] Moved loadUsersByRoleId to state and refactored prepareSelectDropdown
